Pulling PHP out of Clinical reports
